+++ vehicle-syntax/src/Vehicle/Syntax/AST/Instances/NoThunks.hs
@@ -78,7 +78,6 @@ instance NoThunks Expr
 
 -- Vehicle.Syntax.AST.Name
 instance NoThunks Module
-instance NoThunks ModulePath
 instance NoThunks Identifier
 
 -- Vehicle.Syntax.AST.Prog